% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFCF.R
\name{CFCF}
\alias{CFCF}
\title{Causal forest learned (cross-fitted) with training and holdout data, fit
over K folds. Forests are estimated with all but i folds,
then estimates are predicted for the holdout (ith) fold. This is repeated
a specified number of times and the results are averaged across folds.}
\usage{
CFCF(X, Y, W, K = 5, reps = 4, num.trees = 2000)
}
\arguments{
\item{X}{a matrix of covariates, dimension n rows by m columns}

\item{Y}{a vector of outcomes, dimension 1xn}

\item{W}{a vector of treatment status, dimension 1Xn}

\item{K}{the number of folds to be used in cross-fitting, default is 5}

\item{reps}{the number of times to repeat the cross-fitting, default is 4}

\item{num.trees}{the number of trees to be fitted in the causal forest, default is 2000}
}
\value{
A list containing the following components:
  \itemize{
    \item{\code{cates}}{A numeric vector containing the conditional average treatment effects for each unit or group analyzed, averaged over repetitions.}
    \item{\code{errors}}{A numeric vector of errors for each CATE estimation.}
    \item{\code{scores}}{A numeric vector of double-robust scores calculated within the function, according to "get_scores" in the grf package.}
    \item{\code{pscores}}{A numeric vector with the propensity scores estimating the probability of assignment to treatment given covariates.}
    \item{\code{gamma0}}{A numeric value or vector representing the double-robust score for the control group, according to "double_robust_scores" in the grf package.}
    \item{\code{gamma1}}{A numeric value or vector representing the double-robust score for the treated group, according to "double_robust_scores" in the grf package}
    \item{\code{ATE}}{A numeric value indicating the Average Treatment Effect estimated by the function, averaged over repetitions.}
    \item{\code{SE}}{A numeric vector containing the standard errors associated with ethe overall ATE.}
  }

Note that the 'alpha' and 'honesty.fraction' parameters within the causal forest function
will be tuned within the training step, see documentation for grf package
}
\description{
Causal forest learned (cross-fitted) with training and holdout data, fit
over K folds. Forests are estimated with all but i folds,
then estimates are predicted for the holdout (ith) fold. This is repeated
a specified number of times and the results are averaged across folds.
}
